VO-2902: Convert static strings in views to use translate directive
I've tried to reduce the string converts to include the least amount of HTML if possible. The cases that do include HTML are phrases that probably need to stick together. We will have another commit focusing on dynamic text. I want to look deeper into the best solution for that.