VO-2902 - Setup Angular Gettext to pull translations in app
Setup so local dev will be able to see translated strings.
Any strings with missing translation for the specified language will have two underscores wrapped around it (__Example Text__) for easy identifying.
Any strings missing the translation call will have no text wrapping around it. These will need to be modified to be translation wrapped.
When in the default language (en), we still wrap with missing translation, but we don't truly need a translation for english.