VO-2902 - Added translation debugger tool for local use
You will now see a Translations panel on the left of the screen when debugging locally. This is to easily switch without the need of using the real switcher. There is also an "English (Debug)" mode which will highlight all the translated strings.

Some of the Spanish translations have been added as well.